save car images in internal storage

+++ app/src/main/java/com/cleverlycode/getwheels/data/repositories/CarsRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package com.cleverlycode.getwheels.data.repositories
 
+import android.graphics.Bitmap
 import com.cleverlycode.getwheels.data.local.CarsDatabase
 import com.cleverlycode.getwheels.data.mapper.toCar
 import com.cleverlycode.getwheels.data.mapper.toCarDetail
@@ -8,12 +9,8 @@ import com.cleverlycode.getwheels.data.remote.CarsService
 import com.cleverlycode.getwheels.domain.models.Car
 import com.cleverlycode.getwheels.domain.models.CarDetail
 import com.cleverlycode.getwheels.domain.repositories.CarsRepository
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.transform
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class CarsRepositoryImpl @Inject constructor(
@@ -26,52 +23,39 @@ class CarsRepositoryImpl @Inject constructor(
         val cars = dao.observeCars().transform { carEntitiesList ->
             val carsList: MutableList<Car> = mutableListOf()
             carEntitiesList.forEach { carEntity ->
-                runBlocking {
-                    val car = carEntity.toCar()
-                    val imageRef = async(coroutineContext) {
-                        carsService.getCarsPictureRefList(docId = carEntity.id)[0]
-                    }
-                    imageRef.join()
-                    car.imageRef = imageRef.await()
-                    carsList.add(car)
-                }
+                val car = carEntity.toCar()
+                car.imageRef = carsService.getCarsPictureRef(docId = carEntity.id)
+                carsList.add(car)
             }
             emit(carsList)
         }
-
         return cars
     }
 
-    override suspend fun getCarDetails(carId: String): CarDetail? {
-        return dao.getCar(carId)?.toCarDetail()
-    }
+    override suspend fun getCars(): List<Car> =
+        dao.getCars().map { carEntity ->
+            carEntity.toCar()
+        }
+
+    override suspend fun getCarDetails(carId: String): CarDetail? =
+        dao.getCar(carId)?.toCarDetail()
 
     override suspend fun getCarsPictureRefList(carId: String): List<String> {
         TODO("Not yet implemented")
     }
 
     override suspend fun syncWithRemoteDataSource() {
-        withContext(Dispatchers.IO) {
-            val cars = carsService.getCars()
+        val carsResource = carsService.getCars()
+        if (carsResource.data != null) {
+            val cars = carsResource.data
             dao.deleteAllCars()
-            cars.data?.forEach { car ->
+            cars.forEach { car ->
                 dao.insert(carEntity = car.toCarEntity())
             }
         }
     }
 
-    override suspend fun syncWithRemoteDataSource(userId: String) {
-        withContext(Dispatchers.IO) {
-//            val cars = carsService.getCars(userId = userId)
-//
-//            dao.deleteAllCars()
-//            cars.forEach { car ->
-//                dao.insert(carEntity = car.toCarEntity())
-//            }
-        }
-    }
+    override suspend fun getCarImage(carId: String): Bitmap? =
+        carsService.getCarPicture(carId = carId)
 
-    override suspend fun getCarsImages(userId: String) {
-        TODO("Not yet implemented")
-    }
 }
